refactor: 메서드 네이밍 변경 및 패키지 이동
