+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handler/adapter/HandlerMapper.java
@@ -6,5 +6,5 @@ public interface HandlerMapper {
 
     void initialize();
 
-    Object getController(final HttpServletRequest req);
+    Object getHandler(final HttpServletRequest req);
 }
